use better solution to block aliases
I SWEAR TO GOD, I'LL REMOVE THIS COMMAND HIDER!!!

Ok anyway. Previously, we used an array that was converted into a list to check for the presence of 'aliases' or 'single' in that list. While this approach worked, it was not efficient. Although we could block aliases in tab completion, the main command remained accessible.

Now, we have created a boolean variable that is stored in a group. This simplifies the check for whether to block aliases.

And it ACTUALY WORKS FINE! wow